backend: similarity model score field fixed

+++ backend/recommender/migrations/0001_add_similarity_model.py
@@ -19,7 +19,7 @@ class Migration(migrations.Migration):
                 ('id', models.BigAutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
                 ('source_id', models.PositiveIntegerField()),
                 ('target_id', models.PositiveIntegerField()),
-                ('score', models.SmallIntegerField()),
+                ('score', models.DecimalField(decimal_places=2, max_digits=3)),
                 ('content_type', models.ForeignKey(on_delete=django.db.models.deletion.CASCADE, to='contenttypes.contenttype')),
             ],
         ),
